+++ middleware/index.ts
@@ -1,273 +1,22 @@
 import {Logger, LoggerLevel} from '@salesforce/core/lib/logger';
-import {SObject} from '@salesforce/salesforce-sdk/dist/objects';
-import {Constants, APIVersion} from '@salesforce/salesforce-sdk/dist/index';
-import {
-    ConnectionConfig,
-    DataApi,
-    ErrorResult,
-    Secrets,
-    SuccessResult,
-    UnitOfWork,
-    UnitOfWorkGraph,
-} from '@salesforce/salesforce-sdk/dist/api';
-
-import {
-    Context,
-    InvocationEvent,
-    Org,
-    User,
-} from '@salesforce/salesforce-sdk/dist/functions';
 const {Message} = require('@projectriff/message');
 const http = require('http');
 const https = require('https');
+import {applySfFnMiddleware} from './lib/sfMiddleware';
+import {
+    FunctionInvocationRequest, 
+    saveFnInvocation, 
+    saveFnInvocationError
+} from './lib/FunctionInvocationRequest';
+import {
+    ASYNC_CE_TYPE,
+    ASYNC_FULFILL_HEADER,
+    FN_INVOCATION,
+    X_FORWARDED_HOST,
+    X_FORWARDED_PROTO
+} from './lib/constants';
+import loadUserFunction from './userFnLoader';
 
-import loadUserFunction from './userFnLoader'
-
-const ASYNC_FULFULLMENT_HEADER = 'X-Async-Fulfill-Request';
-const FN_INVOCATION = 'fnInvocation';
-
-enum FunctionInvocationRequestStatusEnum {
-    Success = 'Success',
-    Error = 'Error'
-}
-
-// TODO: Remove when FunctionInvocationRequest is deprecated.
-class FunctionInvocationRequest {
-  public response: any;
-  public status: FunctionInvocationRequestStatusEnum;
-
-  constructor(public readonly id: string,
-              private readonly logger: Logger,
-              private readonly dataApi?: DataApi) {
-  }
-
-  /**
-   * Saves FunctionInvocationRequest
-   *
-   * @throws err if response not provided or on failed save
-   */
-  public async save(): Promise<any> {
-      if (!this.response) {
-          throw new Error('Response not provided');
-      }
-
-      if (this.dataApi) {
-          const responseBase64 = Buffer.from(JSON.stringify(this.response)).toString('base64');
-
-          try {
-              // Prime pump (W-6841389)
-              const soql = `SELECT Id, FunctionName, Status, CreatedById, CreatedDate FROM FunctionInvocationRequest WHERE Id ='${this.id}'`;
-              await this.dataApi.query(soql);
-          } catch (err) {
-              this.logger.warn(err.message);
-          }
-
-          const fnInvocation = new SObject('FunctionInvocationRequest').withId(this.id);
-          fnInvocation.setValue('ResponseBody', responseBase64);
-          if (this.status) {
-            fnInvocation.setValue('Status', this.status.toString());
-          }
-          const result: SuccessResult | ErrorResult = await this.dataApi.update(fnInvocation);
-          if (!result.success && 'errors' in result) {
-              // Tells tsc that 'errors' exist and join below is okay
-              const msg = `Failed to send response [${this.id}]: ${result.errors.join(',')}`;
-              this.logger.error(msg);
-              throw new Error(msg);
-          } else {
-              return result;
-          }
-      } else {
-          throw new Error('Authorization not provided');
-      }
-  }
-}
-
-function headersToMap(headers: any = {}): ReadonlyMap<string, ReadonlyArray<string>> {
-    const headersMap: Map<string, ReadonlyArray<string>> = new Map(Object.entries(headers));
-    return headersMap;
-}
-
-/**
- * Construct Event from invocation request.
- *
- * @param data    -- function payload
- * @param headers -- request headers
- * @param payload -- request payload
- * @return event
- */
-function createEvent(data: any, headers: any, payload: any): InvocationEvent {
-    return new InvocationEvent(
-        data,
-        payload.contentType,
-        payload.schemaURL,
-        payload.id,
-        payload.source,
-        payload.time,
-        payload.type,
-        headersToMap(headers)
-    );
-}
-
-/**
- * Construct User object from the request context.
- *
- * @param userContext -- userContext object representing invoking org and user
- * @return user
- */
-function createUser(userContext: any): User {
-    return new User(
-        userContext.userId,
-        userContext.username,
-        userContext.onBehalfOfUserId
-    );
-}
-
-/**
- * Construct Secrets object with logger.
- *
- *
- * @param logger -- logger to use in case of secret load errors
- * @return secrets loader/cache
- */
-function createSecrets(logger: Logger): Secrets {
-    return new Secrets(logger);
-}
-
-/**
- * Construct Org object from the request context.
- *
- * @param reqContext
- * @return org
- */
-function createOrg(logger: Logger, reqContext: any, accessToken?: string): Org {
-    const userContext = reqContext.userContext;
-    if (!userContext) {
-        const message = `UserContext not provided: ${JSON.stringify(reqContext)}`;
-        throw new Error(message);
-    }
-
-    const apiVersion = reqContext.apiVersion || process.env.FX_API_VERSION || Constants.CURRENT_API_VERSION;
-    const user = createUser(userContext);
-
-    // If accessToken was provided, setup APIs.
-    let dataApi: DataApi | undefined;
-    let unitOfWork: UnitOfWork | undefined;
-    let unitOfWorkGraph: UnitOfWorkGraph | undefined;
-    if (accessToken) {
-        const config: ConnectionConfig = new ConnectionConfig(
-            accessToken,
-            apiVersion,
-            userContext.salesforceBaseUrl
-        );
-        unitOfWork = new UnitOfWork(config, logger);
-        if (apiVersion >= APIVersion.V50) {
-            unitOfWorkGraph = new UnitOfWorkGraph(config, logger);
-        }
-        dataApi = new DataApi(config, logger);
-    }
-
-    return new Org(
-        apiVersion,
-        userContext.salesforceBaseUrl,
-        userContext.orgDomainUrl,
-        userContext.orgId,
-        user,
-        dataApi,
-        unitOfWork,
-        unitOfWorkGraph
-    );
-}
-
-/**
- * Construct Context from function payload.
- *
- * @param id                   -- request payload id
- * @param logger               -- logger
- * @param secrets              -- secrets convenience class
- * @param reqContext           -- reqContext from the request, contains salesforce stuff (user reqContext, etc)
- * @param accessToken          -- accessToken for function org access, if provided
- * @param functionInvocationId -- FunctionInvocationRequest ID, if applicable
- * @return context
- */
-function createContext(id: string, logger: Logger, secrets: Secrets, reqContext?: any,
-                       accessToken?: string, functionInvocationId?: string): Context {
-    if (typeof reqContext === 'string') {
-        reqContext = JSON.parse(reqContext);
-    }
-
-    const org = reqContext ? createOrg(logger, reqContext!, accessToken) : undefined;
-    const context = new Context(id, logger, org, secrets);
-
-    // If functionInvocationId is provided, create and set FunctionInvocationRequest object
-    let fnInvocation: FunctionInvocationRequest;
-    if (accessToken && functionInvocationId) {
-        fnInvocation = new FunctionInvocationRequest(functionInvocationId, logger, org.data);
-        context[FN_INVOCATION] = fnInvocation;
-    }
-    return context;
-}
-
-/**
- * Initialize Salesforce SDK for function invocation.
- *
- * @param request     -- contains {payload, headers}
- * @param logger      -- Logger
- * @return returnArgs -- array of arguments that make-up the user functions arguments
- */
-function applySfFnMiddleware(request: any, logger: Logger): Array<any> {
-    // Validate the input request
-    if (!request) {
-        throw new Error('Request Data not provided');
-    }
-
-    //use secret here in lieu of DEBUG runtime environment var until we have deployment time support of config var
-    const secrets = createSecrets(logger);
-    const debugSecret = secrets.getValue('sf-debug', 'DEBUG');
-    logger.info(`DEBUG flag is ${debugSecret ? debugSecret : 'unset'}`);
-    if(debugSecret || LoggerLevel.DEBUG === logger.getLevel() || process.env.DEBUG) {
-        //for dev preview, we log the ENTIRE raw request, may need to filter sensitive properties out later
-        //the hard part of filtering is to know which property name to filter
-        //change the logger level, so any subsequent user function's logger.debug would log as well
-        logger.setLevel(LoggerLevel.DEBUG);
-        logger.debug('debug raw request in middleware');
-        logger.debug(request);
-    }
-
-    const data = request.payload.data;
-    if (!data) {
-        throw new Error('Data field of the cloudEvent not provided in the request');
-    }
-
-    if (!data.context) {
-        logger.warn('Context not provided in data: context is partially initialize');
-    }
-
-    // Not all functions will require an accessToken used for org access.
-    // The accessToken will not be passed directly to functions, but instead
-    // passed as part of SFContext used in SF middleware to setup API instances.
-    let accessToken: string;
-    let functionInvocationId: string;
-    if (data.sfContext) {
-        accessToken = data.sfContext.accessToken || undefined;
-        functionInvocationId = data.sfContext.functionInvocationId || undefined;
-        // Internal only
-        delete data.sfContext;
-    }
-
-    // Construct event contain custom payload and details about the function request
-    const event = createEvent(data.payload, request.headers, request.payload);
-
-    // Construct invocation context, to be sent to the user function.
-    const context = createContext(request.payload.id,
-                                  logger,
-                                  secrets,
-                                  data.context,
-                                  accessToken,
-                                  functionInvocationId);
-
-    // Function params
-    return [event, context, logger];
-}
 
 function createLogger(requestID?: string): Logger {
     const logger = new Logger('Evergreen Logger');
@@ -296,48 +45,23 @@ function errorMessage(error: Error): any {
 }
 
 function isAsyncRequest(type: string) {
-    return type && type.startsWith('com.salesforce.function.async');
-}
-
-// Save update to FunctionInvocationRequest
-// Does not throw
-async function saveFnInvocation(logger: Logger, 
-                                fnInvocation: FunctionInvocationRequest, 
-                                response: any, 
-                                status: FunctionInvocationRequestStatusEnum = FunctionInvocationRequestStatusEnum.Success): Promise<void> {
-    if (!fnInvocation) {
-        return;
-    }
-
-    try {
-        fnInvocation.status = status;
-        fnInvocation.response = response;
-        await fnInvocation.save();
-    } catch (err) {
-        logger.error(`Unable to save function response [${fnInvocation.id}]: ${err.message}`);
-    }
-}
-
-async function saveFnInvocationError(logger: Logger, fnInvocation: FunctionInvocationRequest, response: any): Promise<void> {
-    await saveFnInvocation(logger, fnInvocation, response, FunctionInvocationRequestStatusEnum.Error);
+    return type && type.startsWith(ASYNC_CE_TYPE);
 }
 
 // Invoke given function again to fulfill async request; response is not handled
-async function invokeAsyncFn(logger: Logger, payload: any, headers: any, fnInvocation: FunctionInvocationRequest) {
+async function invokeAsyncFn(logger: Logger, payload: any, headers: any): Promise<void> {
     // host header is required
-    const hostKey = Object.keys(headers).find(k => 'x-forwarded-host' === k.toLowerCase());
+    const hostKey = Object.keys(headers).find(k => X_FORWARDED_HOST === k.toLowerCase());
     const hostUrl = headers[hostKey];
     if (!hostUrl) {
         throw new Error('Unable to determine host for async invocation');
     }
-    delete headers[hostKey];
 
     const hostParts = hostUrl.split(':');
-    const protocolKey = Object.keys(headers).find(k => 'x-forwarded-proto' === k.toLowerCase());
+    const protocolKey = Object.keys(headers).find(k => X_FORWARDED_PROTO === k.toLowerCase());
     let isHttps = true;
     if (headers[protocolKey]) {
         isHttps = headers[protocolKey].toLowerCase() === 'https';
-        delete headers[protocolKey];
     }
 
     const options = {
@@ -345,19 +69,20 @@ async function invokeAsyncFn(logger: Logger, payload: any, headers: any, fnInvoc
         port: hostParts[1] ? parseInt(hostParts[1]) : (isHttps ? 443 : 80),
         method: 'POST',
         path: '/',
-        headers: {...headers, [ASYNC_FULFULLMENT_HEADER]: true}
+        headers: {...headers, [ASYNC_FULFILL_HEADER]: true}
     };
 
     // Invoke function and ignore response
     const lib = isHttps ? https : http;
-    await new Promise((resolve, reject) => {
+    return new Promise((resolve, reject) => {
+        const expectedErrCode = 'EXPECTED_ECONNRESET';
         const req = lib.request(options);
         req.on('error', async (err) => {
-            // Expect ECONNRESET when request was destroyed
-            if (!(req.destroyed && err['code'] === 'ECONNRESET')) {
-                // Save error to associated function invocation request
-                await saveFnInvocationError(logger, fnInvocation, err.message);
-                
+            if (req.destroyed && err['code'] === expectedErrCode) {
+                // Expected to terminate response
+                resolve();
+            } else {
+                // Return error for async request client
                 reject(err);
             }
         });
@@ -370,10 +95,9 @@ async function invokeAsyncFn(logger: Logger, payload: any, headers: any, fnInvoc
             logger.info('Forwarded async request');
 
             // Forget response and destroy the socket
-            req.destroy();
-
-            // Don't care about response
-            resolve();
+            const expectedErr = new Error();
+            expectedErr['code'] = expectedErrCode;
+            req.destroy(expectedErr);
         });
     });
 };
@@ -390,18 +114,13 @@ export default async function systemFn(message: any): Promise<any> {
     const requestId = headers['Ce-Id'] || headers['X-Request-Id'];
     const requestLogger = createLogger(requestId);
     let isAsync = false;
-    let fnInvocation: FunctionInvocationRequest;
     try {
-        // Create function param objects from request
-        const [event, context, logger] = await applySfFnMiddleware({payload, headers}, requestLogger);
-
         // If initial async request, invoke function again and release request
-        isAsync = isAsyncRequest(event.type);
-        fnInvocation = context[FN_INVOCATION];
+        isAsync = isAsyncRequest(payload.type);
         if (isAsync) {
-            if (!headers[ASYNC_FULFULLMENT_HEADER]) {
+            if (!headers[ASYNC_FULFILL_HEADER]) {
                 requestLogger.info('Received initial async request');
-                await invokeAsyncFn(requestLogger, payload, headers, fnInvocation);
+                await invokeAsyncFn(requestLogger, payload, headers);
                 // Function invoked on forwarded request; release initial client request
                 return Message.builder()
                     .addHeader('content-type', 'application/json')
@@ -414,13 +133,18 @@ export default async function systemFn(message: any): Promise<any> {
         }
 
         let result: any;
+        let fnInvocation: FunctionInvocationRequest;
         try {
+            // Create function param objects from request
+            const [event, context, logger] = applySfFnMiddleware({payload, headers}, requestLogger);
+            fnInvocation = context[FN_INVOCATION];
+
             // Invoke requested function
-            result = await userFn([event, context, logger]);
+            result = await userFn(...[event, context, logger]);
 
             // If async, save result to associated function invocation request
             if (isAsync) {
-                await saveFnInvocation(logger, fnInvocation, result);
+                await saveFnInvocation(requestLogger, fnInvocation, result);
             }
 
             // Currently, riff doesn't support undefined or null return values
@@ -428,7 +152,7 @@ export default async function systemFn(message: any): Promise<any> {
         } catch (invokeErr) {
             // If async, save error to associated function invocation request
             if (isAsync) {
-                await saveFnInvocationError(logger, fnInvocation, invokeErr.message);
+                await saveFnInvocationError(requestLogger, fnInvocation, invokeErr.message);
             }
 
             throw invokeErr;
