+++ middleware/lib/constants.ts
@@ -1,0 +1,5 @@
+export const ASYNC_CE_TYPE = 'com.salesforce.function.invoke.async';
+export const ASYNC_FULFILL_HEADER = 'X-Async-Fulfill-Request';
+export const FN_INVOCATION = 'fnInvocation';
+export const X_FORWARDED_HOST = 'x-forwarded-host';
+export const X_FORWARDED_PROTO = 'x-forwarded-proto';
