+++ middleware/tsconfig.json
@@ -19,6 +19,6 @@
     },
     "include": [
         "./*.ts", 
-        "test/unit/**/*.ts",
+        "lib/**/*.ts"
     ]
 }
